requirements.txt
streamlit
requests
Pillow
torch
transformers
numpy